Changed the structure of the solution
