+++ BattleShip/BattleShip.Console/Direction.cs
@@ -1,0 +1,10 @@
+﻿namespace BattleShip.Console
+{
+    enum Direction
+    {
+        Up, 
+	    Left, 
+	    Right, 
+	    Down
+    }
+}
