+++ BattleShip/BattleShip.Console/Move.cs
@@ -1,0 +1,11 @@
+﻿namespace BattleShip.Console
+{
+    class Move
+    {
+    public int XPosition;
+    public int YPosition;
+
+    public Move(int x, int y) { XPosition = x; YPosition = y; }
+    public Move() { XPosition = 0; YPosition = 0; }
+    }
+}
