+++ BattleShip/BattleShip.Console/Result.cs
@@ -1,0 +1,10 @@
+﻿namespace BattleShip.Console
+{
+    enum Result
+    {
+        Missed,
+	    Hit,
+	    Untouched
+
+    }
+}
