+++ BattleShip/BattleShip.Console/Ships/BattleShip.cs
@@ -1,0 +1,11 @@
+﻿namespace BattleShip.Console.Ships
+{
+    class BattleShip : Ship
+    {
+        public BattleShip()
+        {
+            Size = 4;
+            StartPoint = new Position();
+        }
+    }
+}
