+++ BattleShip/BattleShip.Console/Ships/Cruiser.cs
@@ -1,0 +1,11 @@
+﻿namespace BattleShip.Console.Ships
+{
+    class Cruiser : Ship
+    {
+        public Cruiser()
+        {
+            Size = 3;
+            StartPoint = new Position();
+        }
+    }
+}
