+++ BattleShip/BattleShip.Console/Ships/Destroyer.cs
@@ -1,0 +1,11 @@
+﻿namespace BattleShip.Console.Ships
+{
+    class Destroyer : Ship
+    {
+        public Destroyer()
+        {
+            Size = 1;
+            StartPoint = new Position();
+        }
+    }
+}
