+++ BattleShip/BattleShip.Console/Ships/Submarine.cs
@@ -1,0 +1,11 @@
+﻿namespace BattleShip.Console.Ships
+{
+    class Submarine : Ship
+    {
+        public Submarine()
+        {
+            Size = 2;
+            StartPoint = new Position();
+        }
+    }
+}
