+++ BattleShip/BattleShip.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30523.141
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BattleShip.Console", "BattleShip.Console\BattleShip.Console.csproj", "{14AD4793-1B78-496A-976B-6AE72E7A050B}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{14AD4793-1B78-496A-976B-6AE72E7A050B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{14AD4793-1B78-496A-976B-6AE72E7A050B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{14AD4793-1B78-496A-976B-6AE72E7A050B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{14AD4793-1B78-496A-976B-6AE72E7A050B}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {8F351239-330A-45B1-8C44-84B20486302B}
+	EndGlobalSection
+EndGlobal
