--- BattleShipOnCSharp/Direction.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace BattleShipOnCSharp
-{
-    enum Direction
-    {
-        Up, 
-	    Left, 
-	    Right, 
-	    Down
-    }
-}
