--- BattleShipOnCSharp/Position.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace BattleShipOnCSharp
-{
-    class Position
-    {
-		public int XPosition;
-		public int YPosition;
-		public Direction direction;
-
-        public Position()
-        {
-            XPosition = 0;
-            YPosition = 0;
-            direction = Direction.Down;
-        }
-		public Position(int x, int y, Direction drt)
-        {
-            XPosition = x;
-            YPosition = y;
-            direction = drt;
-        }
-	}
-}
