--- BattleShipOnCSharp/Result.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace BattleShipOnCSharp
-{
-    enum Result
-    {
-        Missed,
-	    Hit,
-	    Untouched
-
-    }
-}
