--- BattleShipOnCSharp/Ships/BattleShip.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace BattleShipOnCSharp
-{
-    class BattleShip : Ship
-    {
-        public BattleShip()
-        {
-            Size = 4;
-            StartPoint = new Position();
-        }
-    }
-}
