--- BattleShipOnCSharp/Ships/Destroyer.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace BattleShipOnCSharp
-{
-    class Destroyer : Ship
-    {
-        public Destroyer()
-        {
-            Size = 1;
-            StartPoint = new Position();
-        }
-    }
-}
