--- BattleShipOnCSharp/Ships/Submarine.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace BattleShipOnCSharp
-{
-    class Submarine : Ship
-    {
-        public Submarine()
-        {
-            Size = 2;
-            StartPoint = new Position();
-        }
-    }
-}
